fix(db): introduce versioned migration runner (issue #498)
- Add `schema_version(version, name, applied_at)` table that records
  which migrations have been applied.
- Add `db/migration.rs` with a `Migration` trait and four concrete
  migrations: V1InitialSchema, V2TodoRatingDropColumn,
  V3LogsToExecutionLogs, V4FeishuFkCascade.
- `Database::run_migrations()` reads applied versions and only runs
  unapplied migrations, so steady-state startup cost drops from
  O(all DDL) to O(migrations needed).
- `init_tables()` becomes a thin shell that just calls
  `run_migrations()`.
- Replace 8+ `.ok()` error-swallowing calls with `unwrap_or_else(|e|
  tracing::warn!(...))` so real migration failures surface in logs.
- Expose `get_schema_version()` and `get_applied_migrations()` for
  diagnostics / future /api/schema/version endpoint.
- Add 5 new unit tests covering: fresh-DB schema_version, idempotent
  re-run, version-order invariant, v1 re-run on populated DB, name
  registry match.

Refs issue #498.
